Refactors PDAG into a separate file
