+++ pgmpy/base/__init__.py
@@ -1,6 +1,7 @@
 from .AncestralBase import AncestralBase
-from .DAG import DAG, PDAG
+from .DAG import DAG
 from .MAG import MAG
+from .PDAG import PDAG
 from .UndirectedGraph import UndirectedGraph
 
 __all__ = ["UndirectedGraph", "DAG", "PDAG", "AncestralBase", "MAG"]
